requirements.txt
# Core ML libraries
xgboost>=1.7.0
scikit-learn>=1.3.0
pandas>=1.5.0
numpy>=1.24.0

# FiinQuantX
--extra-index-url https://fiinquant.github.io/fiinquantx/simple
fiinquantx

# Configuration
PyYAML>=6.0

# Optimization
optuna>=3.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0

# Data validation
pydantic>=1.10.0

# Logging
loguru>=0.7.0

# Environment variables
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 